dev_requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
astroid==2.15.2
asttokens==2.2.1
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
bracex==2.3.post1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
cryptography==38.0.4
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
Django==4.1.4
django-allauth==0.51.0
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-js-asset==2.0.0
django-taggit==3.1.0
django_debug_toolbar==3.8.1
docopt==0.6.2
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.2
fqdn==1.5.1
ghp-import==2.1.0
idna==3.4
importlib-metadata==5.2.0
iniconfig==2.0.0
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
isoduration==20.11.0
isort==5.12.0
jedi==0.18.2
Jinja2==3.1.2
Js2Py==0.74
json5==0.9.10
jsonpointer==2.3
jsonschema==4.17.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.5.0
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.6.1
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.17.0
lazy-object-proxy==1.9.0
lxml==4.9.2
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mergedeep==1.3.4
mistune==2.0.4
mkdocs==1.4.2
mkdocs-awesome-pages-plugin==2.8.0
mkdocs-material==9.0.0
mkdocs-material-extensions==1.1.1
natsort==8.2.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
oauthlib==3.2.2
packaging==22.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
pipdeptree==2.3.3
pipwin==0.5.2
platformdirs==2.6.2
pluggy==1.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyjsparser==2.7.1
PyJWT==2.6.0
pylint==2.17.2
pymdown-extensions==9.9
PyPrind==2.11.3
pyrsistent==0.19.3
pySmartDL==1.3.4
pytest==7.2.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
python3-openid==3.2.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywin32==305
pywinpty==2.0.9
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.7
tornado==6.2
traitlets==5.8.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uri-template==1.2.0
urllib3==1.26.13
watchdog==2.2.0
wcmatch==8.4.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
wrapt==1.15.0
zipp==3.11.0
